New UI stylesheets, teacher dashboard, and more
The site-wide stylesheets were switched to Material Design.  Plans are made to switch them to Polymer or Semantic in the future, though for now a simple bootstrap theme will have to do.

Teacher dashboard was also made, allowing for a teacher to enroll and drop students from a class that they have been assigned to.  Grading and assignment systems are still yet to be added.

The home has now been merged with the dashboard.  Sideboard improvements are planned, along with inclusion of sideboards to dashboards across all user roles.

API was reorganized so that functions pertaining to a certain object in the system are classified by that object and not by that who performs it.  This not only simplifies routing, but allows for "if (role isAtleastTeacher)" and similar.